backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "login-admin-backend"
version = "0.1.0"
description = "Backend de autenticacao com bootstrap de admin inicial"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "psycopg[binary,pool]>=3.2",
    "pyjwt[crypto]>=2.9",
    "argon2-cffi>=23.1",
    "pydantic>=2.9",
    "pydantic-settings>=2.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "httpx>=0.27",
    "mypy>=1.11",
    "ruff>=0.6",
    "pytest-cov>=5.0",
    "pytest-benchmark>=4.0",
    "freezegun>=1.5",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.mypy]
python_version = "3.12"
strict = true
files = ["src"]
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_unreachable = true
show_error_codes = true
namespace_packages = true
explicit_package_bases = true

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "ANN"]
ignore = ["ANN101", "ANN102"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = "--strict-markers --maxfail=1 -q"
xfail_strict = true
filterwarnings = [
    "error::DeprecationWarning",
]
markers = [
    "unit: unit tests (no external dependencies)",
    "contract: HTTP contract tests using TestClient",
    "integration: integration tests requiring PostgreSQL",
    "performance: performance/latency benchmark tests",
]
